+++ packages/tdesign-vue-next-chat/site/site.config.mjs
@@ -68,20 +68,20 @@ export const docs = [
         titleEn: 'Getting Started',
         name: 'getting-started',
         path: '/chat/getting-started',
-        component: () => import('@/packages/tdesign-vue-next-chat/site/docs/getting-started.md'),
+        component: () => import('./docs/getting-started.md'),
       },
       {
         title: '更新日志',
         titleEn: 'CHANGELOG',
         name: 'changelog',
         path: '/chat/changelog',
-        component: () => import('@/packages/tdesign-vue-next-chat/CHANGELOG.md'),
+        component: () => import('../CHANGELOG.md'),
       },
       {
         title: '什么是流式输出',
         name: 'sse',
         path: '/chat/sse',
-        component: () => import('@/packages/tdesign-vue-next-chat/site/docs/sse.md'),
+        component: () => import('./docs/sse.md'),
       },
     ],
   },
@@ -95,16 +95,16 @@ export const docs = [
         titleEn: 'Theme Customization',
         name: 'custom-theme',
         path: '/chat/custom-theme',
-        component: () => import('@common/theme.md'),
-        componentEn: () => import('@common/theme.en-US.md'),
+        component: () => import('@tdesign/common/theme.md'),
+        componentEn: () => import('@tdesign/common/theme.en-US.md'),
       },
       {
         title: '暗色模式',
         titleEn: 'Dark Mode',
         name: 'dark-mode',
         path: '/chat/dark-mode',
-        component: () => import('@common/dark-mode.md'),
-        componentEn: () => import('@common/dark-mode.en-US.md'),
+        component: () => import('@tdesign/common/dark-mode.md'),
+        componentEn: () => import('@tdesign/common/dark-mode.en-US.md'),
       },
     ],
   },
