+++ packages/tdesign-vue-next-chat/site/src/main.js
@@ -10,7 +10,7 @@ import BaseUsage from './components/base-usage.vue';
 
 // import tdesign style
 import '@tdesign/components/style/index.js';
-import '@common/style/web/docs.less';
+import '@tdesign/common/style/web/docs.less';
 
 // import site webComponents
 import 'tdesign-site-components';
